# apps/backend/openapi-quotes.yaml
openapi: 3.0.3
info:
  title: Pivotal Flow Quotes API
  description: |
    API for managing quotes in the Pivotal Flow system.
    
    ## Features
    - Create, read, update, and delete quotes
    - Quote status transitions (draft → pending → approved → sent → accepted)
    - Automatic calculations (subtotal, tax, discounts, total)
    - Multi-tenant support with organization isolation
    - Audit logging for all quote changes
    - Quote number generation with organization prefixes
    
    ## Authentication
    All endpoints require authentication via JWT token in the Authorization header.
    The token must include organization and user context for multi-tenant access.
    
    ## Multi-tenancy
    All quotes are scoped to the user's organization. Users can only access quotes
    within their own organization.
  version: 1.0.0
  contact:
    name: Pivotal Flow Team
    email: support@pivotalflow.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.pivotalflow.com
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing user and organization context.
        Token must include:
        - `sub`: User ID
        - `org`: Organization ID
        - `iat`: Issued at timestamp
        - `exp`: Expiration timestamp
  schemas:
    QuoteStatus:
      type: string
      enum: [draft, pending, approved, sent, accepted, rejected, cancelled]
      description: Current status of the quote
      example: "draft"
    QuoteType:
      type: string
      enum: [project, service, product, recurring, one_time]
      description: Type of quote
      example: "project"
    LineItemType:
      type: string
      enum: [service, product, material, travel, expense, discount, tax]
      description: Type of line item
      example: "service"
    DiscountType:
      type: string
      enum: [percentage, fixed_amount, per_unit]
      description: Type of discount
      example: "percentage"
    Money:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: string
          description: Amount as a decimal string for precision
          example: "100.00"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO 4217 currency code
          example: "NZD"
    LineItem:
      type: object
      required:
        - lineNumber
        - description
        - quantity
        - unitPrice
      properties:
        lineNumber:
          type: integer
          minimum: 1
          description: Sequential line number
          example: 1
        type:
          $ref: '#/components/schemas/LineItemType'
        description:
          type: string
          minLength: 1
          maxLength: 1000
          description: Item description
          example: "Web Development Services"
        quantity:
          type: number
          minimum: 0
          description: Quantity of items
          example: 40
        unitPrice:
          $ref: '#/components/schemas/Money'
        unitCost:
          $ref: '#/components/schemas/Money'
        taxRate:
          type: number
          minimum: 0
          maximum: 1
          description: Tax rate as decimal (0.15 = 15%)"
          example: 0.15
        discountType:
          $ref: '#/components/schemas/DiscountType'
        discountValue:
          type: number
          minimum: 0
          description: Discount value
          example: 0
        serviceCategoryId:
          type: string
          format: uuid
          description: Service category ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        rateCardId:
          type: string
          format: uuid
          description: Rate card ID for pricing
          example: "550e8400-e29b-41d4-a716-446655440001"
        metadata:
          type: object
          description: Additional metadata
          example: { "projectPhase": "development", "priority": "high" }
    Quote:
      type: object
      required:
        - id
        - quoteNumber
        - customerId
        - title
        - status
        - type
        - validFrom
        - validUntil
        - currency
        - subtotal
        - taxAmount
        - totalAmount
        - createdBy
        - organizationId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique quote identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        quoteNumber:
          type: string
          description: Human-readable quote number with organization prefix
          example: "Q-2025-0001"
        customerId:
          type: string
          format: uuid
          description: Customer ID
          example: "550e8400-e29b-41d4-a716-446655440001"
        projectId:
          type: string
          format: uuid
          description: Associated project ID
          example: "550e8400-e29b-41d4-a716-446655440002"
        title:
          type: string
          minLength: 3
          maxLength: 255
          description: Quote title
          example: "Website Development Project"
        description:
          type: string
          maxLength: 2000
          description: Quote description
          example: "Complete website development including design, development, and deployment"
        status:
          $ref: '#/components/schemas/QuoteStatus'
        type:
          $ref: '#/components/schemas/QuoteType'
        validFrom:
          type: string
          format: date-time
          description: Quote validity start date
          example: "2025-01-01T00:00:00Z"
        validUntil:
          type: string
          format: date-time
          description: Quote validity end date
          example: "2025-12-31T23:59:59Z"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: Quote currency
          example: "NZD"
        exchangeRate:
          type: number
          minimum: 0
          description: Exchange rate to base currency
          example: 1.0
        subtotal:
          type: string
          description: Subtotal amount as decimal string
          example: "8000.00"
        taxRate:
          type: number
          minimum: 0
          maximum: 1
          description: Quote-level tax rate
          example: 0.15
        taxAmount:
          type: string
          description: Total tax amount as decimal string
          example: "1200.00"
        discountType:
          $ref: '#/components/schemas/DiscountType'
        discountValue:
          type: number
          minimum: 0
          description: Quote-level discount value
          example: 0
        discountAmount:
          type: string
          description: Total discount amount as decimal string
          example: "0.00"
        totalAmount:
          type: string
          description: Total amount as decimal string
          example: "9200.00"
        termsConditions:
          type: string
          maxLength: 5000
          description: Terms and conditions
          example: "Payment terms: 50% upfront, 50% on completion"
        notes:
          type: string
          maxLength: 2000
          description: Customer-facing notes
          example: "Includes 3 months of support after launch"
        internalNotes:
          type: string
          maxLength: 2000
          description: Internal notes (not visible to customer)
          example: "Customer prefers React over Vue"
        createdBy:
          type: string
          format: uuid
          description: User who created the quote
          example: "550e8400-e29b-41d4-a716-446655440003"
        approvedBy:
          type: string
          format: uuid
          description: User who approved the quote
          example: "550e8400-e29b-41d4-a716-446655440004"
        approvedAt:
          type: string
          format: date-time
          description: When the quote was approved
          example: "2025-01-15T10:30:00Z"
        sentAt:
          type: string
          format: date-time
          description: When the quote was sent to customer
          example: "2025-01-16T14:00:00Z"
        acceptedAt:
          type: string
          format: date-time
          description: When the quote was accepted by customer
          example: "2025-01-20T09:15:00Z"
        expiresAt:
          type: string
          format: date-time
          description: When the quote expires
          example: "2025-02-01T23:59:59Z"
        organizationId:
          type: string
          format: uuid
          description: Organization ID for multi-tenancy
          example: "550e8400-e29b-41d4-a716-446655440005"
        metadata:
          type: object
          description: Additional metadata
          example: { "source": "sales_team", "priority": "high" }
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2025-01-01T08:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-01-15T10:30:00Z"
        lineItems:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/LineItem'
              - type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    description: Line item ID
                    example: "550e8400-e29b-41d4-a716-446655440006"
                  subtotal:
                    type: string
                    description: Line item subtotal
                    example: "6000.00"
                  taxAmount:
                    type: string
                    description: Line item tax amount
                    example: "900.00"
                  discountAmount:
                    type: string
                    description: Line item discount amount
                    example: "0.00"
                  totalAmount:
                    type: string
                    description: Line item total amount
                    example: "6900.00"
                  createdAt:
                    type: string
                    format: date-time
                    description: Creation timestamp
                    example: "2025-01-01T08:00:00Z"
                  updatedAt:
                    type: string
                    format: date-time
                    description: Last update timestamp
                    example: "2025-01-01T08:00:00Z"
...
    StatusTransitionRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/QuoteStatus'
          description: New status to transition to
          example: "pending"
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: "ValidationError"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request data"
        code:
          type: string
          description: Error code for programmatic handling
          example: "INVALID_INPUT"
        details:
          type: object
          description: Additional error details
          example:
            field: "title"
            value: ""
            constraint: "minLength"
paths:
...
  /v1/quotes/{id}/status:
    post:
      summary: Transition quote status
      description: |
        Transition a quote to a new status.
        
        ## Status Workflow
        Quotes follow a specific workflow:
        1. `draft` - Initial state, can be edited
        2. `pending` - Submitted for approval
        3. `approved` - Approved by manager
        4. `sent` - Sent to customer
        5. `accepted` - Accepted by customer
        6. `rejected` - Rejected by customer or manager
        7. `cancelled` - Cancelled (can be done from any state)
        
        ## Validation
        - Only valid status transitions are allowed
        - Some transitions require specific conditions (e.g., approval requires manager role)
        - Audit logging records all status changes
        
        ## Automatic Actions
        - `approved`: Sets `approvedBy` and `approvedAt`
        - `sent`: Sets `sentAt`
        - `accepted`: Sets `acceptedAt`
      tags:
        - Quotes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Quote ID
          example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusTransitionRequest'
            examples:
              to_pending:
                summary: Transition to pending
                value:
                  status: "pending"
              to_approved:
                summary: Transition to approved
                value:
                  status: "approved"
              to_sent:
                summary: Transition to sent
                value:
                  status: "sent"
              to_accepted:
                summary: Transition to accepted
                value:
                  status: "accepted"
      responses:
        '200':
          description: Status transition successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
              examples:
                status_updated:
                  summary: Status updated response
                  value:
                    id: "550e8400-e29b-41d4-a716-446655440000"
                    quoteNumber: "Q-2025-0001"
                    customerId: "550e8400-e29b-41d4-a716-446655440001"
                    title: "Website Development Project"
                    status: "approved"
                    type: "project"
                    validFrom: "2025-01-01T00:00:00Z"
                    validUntil: "2025-12-31T23:59:59Z"
                    currency: "NZD"
                    exchangeRate: 1.0
                    subtotal: "8000.00"
                    taxRate: 0.15
                    taxAmount: "1200.00"
                    discountType: "percentage"
                    discountValue: 0
                    discountAmount: "0.00"
                    totalAmount: "9200.00"
                    createdBy: "550e8400-e29b-41d4-a716-446655440003"
                    approvedBy: "550e8400-e29b-41d4-a716-446655440004"
                    approvedAt: "2025-01-15T10:30:00Z"
                    organizationId: "550e8400-e29b-41d4-a716-446655440005"
                    createdAt: "2025-01-01T08:00:00Z"
                    updatedAt: "2025-01-15T10:30:00Z"
                    lineItems: []
        '400':
          description: Bad request - invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_transition:
                  summary: Invalid status transition
                  value:
                    error: "InvalidStatusTransition"
                    message: "Cannot transition from 'draft' to 'accepted'"
                    code: "INVALID_STATUS_TRANSITION"
                    details:
                      fromStatus: "draft"
                      toStatus: "accepted"
                      allowedTransitions: ["pending", "cancelled"]
        '404':
          description: Quote not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - quote cannot be transitioned in current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions for status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...